Disable interspersed args in the python script
This will make the parser stop parsing args when the first non-arg
element is hit (in this case the parser).

Therefore, doing something like `python2 $PATH_TO_SCRIPT compile -p
root` will work out of the box without the need of `--`, which is
confusing.

We could have added `--` to the docs, but seriously: people will get
this wrong and I prefer to fix this upstream than educating my users.

Fixes scalacenter/bloop#100.