Find blp-coursier in directory where pynailgun is run
Bloop can be installed in different places in a machine. For example, in
macOS bloop is installed under a specific Homebrew Cellar directory and
`$HOME/.bloop` does not even exist. This holds for other
distributions/ways of installing the server.

Therefore, the plugin should locate blp-coursier in the same directory
where pynailgun is run (the same way we do it to locate the bloop
server) rather than looking for it in the home directory.

/cc @torkelrogstad